consolidate groupselector code to new component file
